+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* Make Ens#network and Ens#url public
 * Change default ENS source protocol from wss (websocket) to https
 * Make Ens `web3`, `ensContract` and `registrarContract` private properties
 * Ability to provide ENS network configuration as string like `mainnet`, `testnet` etc.